+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Add `AngelEnvironment` class.
     * Add `Angel.environment`.
     * Deprecated `app.isProduction` in favor of `app.environment.isProduction`.
+* Allow setting of `bodyAsObject`, `bodyAsMap`, or `bodyAsList` **exactly once**.
 
 # 2.0.0-alpha.24
 * Add `AngelEnv` class to `core`.